Fix gui scale handling on 1.20.5
On 1.20.5, the change callback of the guiScale option calls
MinecraftClient#onResolutionChanged, which in turn causes the current
screen to be reinitialized. However, UScreen calls updateGuiScale when
the screen is initialized, causing the gui scale to be set back to
the overridden gui scale. This doesn't recurse forever, since Minecraft
only calls option change callbacks when the value changes.

We fix this by setting a field when we are restoring the GUI scale and
bypassing updateGuiScale when this field is set.